+++ crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/OnrampPresenterCoordinator.kt
@@ -202,7 +202,8 @@ internal class OnrampPresenterCoordinator @Inject constructor(
             is PaymentMethodSelection.CardAndBankAccount -> {
                 linkPresenter.presentPaymentMethodsForOnramp(
                     email = clientEmail(),
-                    paymentMethodTypes = selection.type.toLinkType()
+                    paymentMethodTypes = selection.type.toLinkType(),
+                    collectName = selection.type.requiresNameCollection(),
                 )
             }
             is PaymentMethodSelection.GooglePay -> {
@@ -389,3 +390,6 @@ private fun PaymentMethodType.toLinkType(): List<LinkController.PaymentMethodTyp
         PaymentMethodType.CardAndBankAccount -> null
         PaymentMethodType.GooglePay -> error("Google Pay is not supported in LinkController")
     }
+
+private fun PaymentMethodType.requiresNameCollection(): Boolean =
+    this == PaymentMethodType.BankAccount || this == PaymentMethodType.CardAndBankAccount