+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # CHANGELOG
 
-NEXT_VERSION_BUMP: PATCH
+NEXT_VERSION_BUMP: MINOR
 ## XX.XX.XX - 20XX-XX-XX
 
 ## 23.13.1 - 2026-07-23
 
 ### Payments
 * [FIXED][13544](https://github.com/stripe/stripe-android/pull/13544) Fixed an issue where the SDK could fail to correctly reconcile and close out an Alipay payment in test mode.
+
+### PaymentSheet
+* [ADDED] Added `billingDetailsCollectionConfiguration` to `LinkController.Configuration` (private preview).
 
 ## 23.13.0 - 2026-07-20